Replace major operations with InvokeDynamic
This replaces many of the bytecode instructions with InvokeDynamic
instructions that call into a bootstrapper.

Right now, the bootstrapper simply replaces each call site with a constant
call site that invokes ScriptRuntime or OptRuntime just like we did
before.

However, now we have the option to add more sophisticated dynamic link
optimizations. Plus, the bytecode is easier to read and is starting
to be smaller.